+++ highs/ipm/hipo/factorhighs/DenseFactKernel.cpp
@@ -1,6 +1,6 @@
-#include "DenseFact.h"
 #include "CallAndTimeBlas.h"
 #include "DataCollector.h"
+#include "DenseFact.h"
 #include "FactorHiGHSSettings.h"
 #include "ReturnValues.h"
 #include "Swaps.h"
@@ -12,15 +12,16 @@ namespace hipo {
 
 // Dense Factorisation kernel
 
-static std::pair<Int, double> maxInCol(Int j, Int n, Int m, double* A, Int lda) {
+static std::pair<Int64, double> maxInCol(Int64 j, Int64 n, Int64 m, double* A,
+                                         Int64 lda) {
   // Given the symemtric matrix A, of size nxn, accessed with leading dimension
   // lda, in upper triangular format, ignoring rows 0:j-1, find the maximum in
   // row/col m.
 
   double maxval = -1.0;
-  Int r = -1;
+  Int64 r = -1;
   // scan column m, rows j:m-1
-  for (Int row = j; row < m; ++row) {
+  for (Int64 row = j; row < m; ++row) {
     double val = std::abs(A[row + lda * m]);
     if (val > maxval) {
       maxval = val;
@@ -28,7 +29,7 @@ static std::pair<Int, double> maxInCol(Int j, Int n, Int m, double* A, Int lda) 
     }
   }
   // scan row m, columns m+1:n
-  for (Int col = m + 1; col < n; ++col) {
+  for (Int64 col = m + 1; col < n; ++col) {
     double val = std::abs(A[m + lda * col]);
     if (val > maxval) {
       maxval = val;
@@ -38,7 +39,8 @@ static std::pair<Int, double> maxInCol(Int j, Int n, Int m, double* A, Int lda) 
   return {r, maxval};
 }
 
-static void staticReg(double& pivot, Int sign, const Regul& regval, double& totalreg) {
+static void staticReg(double& pivot, Int64 sign, const Regul& regval,
+                      double& totalreg) {
   // apply static regularisation
 
   double old_pivot = pivot;
@@ -49,8 +51,9 @@ static void staticReg(double& pivot, Int sign, const Regul& regval, double& tota
   totalreg = pivot - old_pivot;
 }
 
-static bool blockBunchKaufman(Int j, Int n, double* A, Int lda, Int* swaps, Int* sign,
-                              double thresh, const Regul& regval, double* totalreg,
+static bool blockBunchKaufman(Int64 j, Int64 n, double* A, Int64 lda,
+                              Int64* swaps, Int64* sign, double thresh,
+                              const Regul& regval, double* totalreg,
                               DataCollector& data) {
   // Perform Bunch-Kaufman pivoting within a block of the supernode (see Schenk,
   // Gartner, ETNA 2006).
@@ -65,9 +68,9 @@ static bool blockBunchKaufman(Int j, Int n, double* A, Int lda, Int* swaps, Int*
   bool flag_2x2 = false;
 
   // Find largest diagonal entry in the residual part of the block
-  Int ind_max_diag = -1;
+  Int64 ind_max_diag = -1;
   double max_diag = -1.0;
-  for (Int i = j; i < n; ++i) {
+  for (Int64 i = j; i < n; ++i) {
     double val = std::abs(A[i + lda * i]);
     if (val > max_diag) {
       max_diag = val;
@@ -82,7 +85,7 @@ static bool blockBunchKaufman(Int j, Int n, double* A, Int lda, Int* swaps, Int*
   // Max in column j of diagonal block
   auto res = maxInCol(j, n, j, A, lda);
   double gamma_j = res.second;
-  Int r = res.first;
+  Int64 r = res.first;
   double Ajj = sign[j] > 0 ? A[j + lda * j] + regval.dual
                            : A[j + lda * j] - regval.primal;
 
@@ -148,9 +151,9 @@ static bool blockBunchKaufman(Int j, Int n, double* A, Int lda, Int* swaps, Int*
   return flag_2x2;
 }
 
-Int denseFactK(char uplo, Int n, double* A, Int lda, Int* pivot_sign,
-               double thresh, const Regul& regval, double* totalreg, Int* swaps,
-               double* pivot_2x2, DataCollector& data) {
+Int64 denseFactK(char uplo, Int64 n, double* A, Int64 lda, Int64* pivot_sign,
+                 double thresh, const Regul& regval, double* totalreg,
+                 Int64* swaps, double* pivot_2x2, DataCollector& data) {
   // ===========================================================================
   // Factorisation kernel
   // Matrix A is in format F
@@ -173,15 +176,15 @@ Int denseFactK(char uplo, Int n, double* A, Int lda, Int* pivot_sign,
     if (!swaps || !pivot_2x2) return kRetInvalidInput;
 
     // initialise order of pivots
-    for (Int i = 0; i < n; ++i) swaps[i] = i;
+    for (Int64 i = 0; i < n; ++i) swaps[i] = i;
 
     // allocate space for copy of col(s)
     std::vector<double> temp(n - 1);
     std::vector<double> temp2(n - 1);
 
-    Int step = 1;
+    Int64 step = 1;
 
-    for (Int j = 0; j < n; j += step) {
+    for (Int64 j = 0; j < n; j += step) {
       bool flag_2x2 = false;
 
 #ifdef HIPO_PIVOTING
@@ -213,7 +216,7 @@ Int denseFactK(char uplo, Int n, double* A, Int lda, Int* pivot_sign,
         // save reciprocal of pivot
         A[j + lda * j] = 1.0 / Ajj;
 
-        const Int M = n - j - 1;
+        const Int64 M = n - j - 1;
         if (M > 0) {
           // make copy of row
           callAndTime_dcopy(M, &A[j + (j + 1) * lda], lda, temp.data(), 1,
@@ -251,7 +254,7 @@ Int denseFactK(char uplo, Int n, double* A, Int lda, Int* pivot_sign,
         A[j + 1 + lda * (j + 1)] = i_d2;
         pivot_2x2[j] = i_off;
 
-        const Int M = n - j - 2;
+        const Int64 M = n - j - 2;
         if (M > 0) {
           double* r1 = &A[j + (j + 2) * lda];
           double* r2 = &A[j + 1 + (j + 2) * lda];
