+++ highs/ipm/hipo/factorhighs/FormatHandler.cpp
@@ -6,7 +6,7 @@
 
 namespace hipo {
 
-FormatHandler::FormatHandler(const Symbolic& S, Int sn, const Regul& regul,
+FormatHandler::FormatHandler(const Symbolic& S, Int64 sn, const Regul& regul,
                              std::vector<double>& frontal)
     : S_{&S},
       regul_{regul},
@@ -23,7 +23,7 @@ FormatHandler::FormatHandler(const Symbolic& S, Int sn, const Regul& regul,
 
 void FormatHandler::terminate(std::vector<double>& clique,
                               std::vector<double>& total_reg,
-                              std::vector<Int>& swaps,
+                              std::vector<Int64>& swaps,
                               std::vector<double>& pivot_2x2) {
   // Move local copies of data into their final position.
   // In this way, the shared objects sn_columns_ and schur_contribution_ are
@@ -35,7 +35,7 @@ void FormatHandler::terminate(std::vector<double>& clique,
   pivot_2x2 = std::move(pivot_2x2_);
 
   // Move local regularisation into total regularisation.
-  for (Int i = 0; i < sn_size_; ++i)
+  for (Int64 i = 0; i < sn_size_; ++i)
     total_reg[S_->snStart(sn_) + i] = local_reg_[i];
 
   // This function should not require a lock, since all threads access different
